Added some wip stuff for graphics.

+++ src/custom_object.hpp
@@ -413,8 +413,8 @@ private:
 	game_logic::formula_variable_storage_ptr vars_, tmp_vars_;
 	game_logic::map_formula_callable_ptr tags_;
 
-	variant& get_property_data(int slot) { if(property_data_.size() <= slot) { property_data_.resize(slot+1); } return property_data_[slot]; }
-	variant get_property_data(int slot) const { if(property_data_.size() <= slot) { return variant(); } return property_data_[slot]; }
+	variant& get_property_data(int slot) { if(property_data_.size() <= size_t(slot)) { property_data_.resize(slot+1); } return property_data_[slot]; }
+	variant get_property_data(int slot) const { if(property_data_.size() <= size_t(slot)) { return variant(); } return property_data_[slot]; }
 	std::vector<variant> property_data_;
 	mutable int active_property_;
